% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapa_base}
\alias{mapa_base}
\title{mapa_base}
\format{
sf map of Ecuador by provinces.
}
\usage{
mapa_base
}
\description{
Tibble de aplicaciones SAAS con sus nombres
}
\details{
Contains sf map of Ecuador by provinces.
}
\keyword{mapa_base}
